move json-related conversions into a separate file
