+++ lib/api/src/conversions/mod.rs
@@ -1,2 +1,3 @@
 pub mod inference;
+pub mod json;
 pub mod vectors;
